pubsub-task-status/requirement.txt

functions-framework
google-cloud-pubsub
